Add files via upload
The purpose of this Python notebook code is to clean up and rearrange temperature data gathered from weather stations. A CSV file called "Daily_MODIS_LST_Stations_2010_2025.csv," which includes daily land surface temperature values from several stations between 2010 and 2025, is first requested to be uploaded by the user. After the file has been uploaded, the code analyzes it and processes the data by extracting the geographic coordinates that are kept in a specific text field. This allows the location data to be used for additional research.

After that, the script concentrates on narrowing the dataset down to just 13 distinct weather stations: Abe, Bol, Nav, Ho, Wen, Wa, Sun, Akuse, Kra, Yendi, Fada, Ouaga, and Ouahigouya, in a predetermined order. It converts the data into a wide format from a long format, in which each row represents a single temperature reading for a station on a specific date. This new arrangement makes the data easier to use for time-series analysis or visualization because each row represents a date and each column contains the temperature data for one of the chosen stations.

Once the data has been rearranged, the code stores the new table as a CSV file named rearranged_LST_data.csv and provides a direct download option. The size of the original and modified datasets, a preview of the first few rows, the column names, the entire date range covered by the data, and the number of missing temperature readings per station are all helpful summaries that it prints. In conclusion, this notebook automatically cleans, filters, and reorganizes temperature data, converting unstructured station records into a table that is suitable for analysis.